refactor: change test structure folder
